Reduce package footprint
